+++ src/ui_basic/editbox.cc
@@ -367,11 +367,11 @@ void EditBox::draw(RenderTarget& dst) {
 		// right edge
 		dst.brighten_rect(Recti(get_w() - 2, 0, 2, get_h() - 2), BUTTON_EDGE_BRIGHT_FACTOR);
 		// top edge
-		dst.fill_rect(Rectf(0.f, 0.f, get_w() - 1, 1), black);
-		dst.fill_rect(Rectf(0.f, 1.f, get_w() - 2, 1), black);
+		dst.fill_rect(Recti(0, 0, get_w() - 1, 1), black);
+		dst.fill_rect(Recti(0, 1, get_w() - 2, 1), black);
 		// left edge
-		dst.fill_rect(Rectf(0.f, 0.f, 1, get_h() - 1), black);
-		dst.fill_rect(Rectf(1.f, 0.f, 1, get_h() - 2), black);
+		dst.fill_rect(Recti(0, 0, 1, get_h() - 1), black);
+		dst.fill_rect(Recti(1, 0, 1, get_h() - 2), black);
 	}
 
 	if (has_focus()) {
